// frontend/src/api/generated/api-client.ts
export class CategoryDto implements ICategoryDto {
    id?: number;
    name?: string | undefined;

    constructor(data?: ICategoryDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
        }
    }

    static fromJS(data: any): CategoryDto {
        data = typeof data === 'object' ? data : {};
        let result = new CategoryDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        return data;
    }
}

export interface ICategoryDto {
    id?: number;
    name?: string | undefined;
}

export class ClientDto implements IClientDto {
    id?: number;
    name?: string | undefined;

    constructor(data?: IClientDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
        }
    }

    static fromJS(data: any): ClientDto {
        data = typeof data === 'object' ? data : {};
        let result = new ClientDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        return data;
    }
}

export interface IClientDto {
    id?: number;
    name?: string | undefined;
}

export class ProductDto implements IProductDto {
    id?: number;
    name?: string | undefined;

    constructor(data?: IProductDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
        }
    }

    static fromJS(data: any): ProductDto {
        data = typeof data === 'object' ? data : {};
        let result = new ProductDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        return data;
    }
}

export interface IProductDto {
    id?: number;
    name?: string | undefined;
}

export class UpdateQuantityDto implements IUpdateQuantityDto {
    quantity?: number;

    constructor(data?: IUpdateQuantityDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.quantity = _data["quantity"];
        }
    }

    static fromJS(data: any): UpdateQuantityDto {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateQuantityDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["quantity"] = this.quantity;
        return data;
    }
}

export interface IUpdateQuantityDto {
    quantity?: number;
}
// ...
export class WarehouseRowDto implements IWarehouseRowDto {
    clientId?: number;
    clientName?: string | undefined;
    productId?: number;
    productName?: string | undefined;
    categories?: string[] | undefined;
    quantity?: number;

    constructor(data?: IWarehouseRowDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.clientId = _data["clientId"];
            this.clientName = _data["clientName"];
            this.productId = _data["productId"];
            this.productName = _data["productName"];
            if (Array.isArray(_data["categories"])) {
                this.categories = [] as any;
                for (let item of _data["categories"])
                    this.categories!.push(item);
            }
            this.quantity = _data["quantity"];
        }
    }

    static fromJS(data: any): WarehouseRowDto {
        data = typeof data === 'object' ? data : {};
        let result = new WarehouseRowDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["clientId"] = this.clientId;
        data["clientName"] = this.clientName;
        data["productId"] = this.productId;
        data["productName"] = this.productName;
        if (Array.isArray(this.categories)) {
            data["categories"] = [];
            for (let item of this.categories)
                data["categories"].push(item);
        }
        data["quantity"] = this.quantity;
        return data;
    }
}

export interface IWarehouseRowDto {
    clientId?: number;
    clientName?: string | undefined;
    productId?: number;
    productName?: string | undefined;
    categories?: string[] | undefined;
    quantity?: number;
}